move: ClearIconButton 패키지 이동
